eso_names_package/requirements.txt
attr==0.3.2
ConfigParser==6.0.0
contextlib2==21.6.0
cryptography==41.0.7
Cython==3.0.8
dl==0.1.0
docutils==0.20.1
HTMLParser==0.0.2
ipython==8.20.0
ipywidgets==8.1.1
Jinja2==3.1.3
jnius==1.1.0
keyring==24.3.0
lockfile==0.12.2
mock==5.1.0
numpy==1.26.3
pandas==2.2.0
Pillow==10.2.0
protobuf==4.25.2
pyOpenSSL==23.3.0
python_Levenshtein==0.23.0
railroad==0.5.0
Sphinx==7.2.6
thread==0.1.3
torch==2.1.2
tornado==6.4
trove_classifiers==2024.1.8
truststore==0.8.0
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
